requirements.txt
fastapi
uvicorn
httpx
python-jose
pyjwt
